requirements.txt
aiogram==3.27.0
python-dotenv==1.0.0
pandas==2.3.0
openai==1.55.0
payos==1.1.0
fastapi==0.109.0
uvicorn==0.27.0
qrcode[pil]==7.4.2
pillow>=10.0.0
httpx==0.25.2
psycopg2-binary